requirements.txt
appdirs==1.4.3
distlib==0.3.0
Django==2.2.12
django-filter==2.4.0
filelock==3.0.12
pbr==5.4.5
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
